feat: .env and logs to .dockerignore
Add .env and logs to .dockerignore to exclude them from Docker builds and deployments.

feat:  is_deleted column to roles table

Add the is_deleted column to the roles table in the database schema.

feat: implement softDelete method in RoleRepository

Implement the softDelete method in the RoleRepository to toggle the isDeleted flag of a role.

feat: implement softDelete method in RolesService

Implement the softDelete method in the RolesService to toggle the isDeleted flag of a role.

chore: update config roles in progress